requirements.txt
requests==2.31.0
openai==1.55.3
httpx==0.27.2
GitPython==3.1.43
mcp==1.12.4
pydantic<3,>=2.8.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
structlog==23.1.0
click==8.1.7
fastapi>=0.104.1,<0.114.0          # loosened to allow newer versions compatible with anyio >=4.0
uvicorn==0.24.0
anyio>=3.7.1,<5.0.0  